requirements.txt
openai==1.16.0
llama-index==0.10.20
sentence-transformers==2.7.0
python-dotenv
pandas
python-telegram-bot==20.6